Feature. Controllers for database (#48)
feature/controllers-for-database

### What's done:
  * Reworked to preprocessor
  * Added tests
  * Added liquibase xml files
  * Added entities

+++ save-common/src/commonMain/kotlin/org/cqfn/save/entities/Project.kt
@@ -15,5 +15,5 @@ class Project(
     var type: String,
     var url: String,
     var description: String?,
-    @Id @GeneratedValue var id: Int? = null,
+    @Id @GeneratedValue var id: Long? = null,
 )
